+++ chirho/dynamical/handlers/solver.py
@@ -1,10 +1,4 @@
-import pyro
-
-
-class Solver(pyro.poutine.messenger.Messenger):
-    def _pyro_simulate(self, msg) -> None:
-        # Overwrite the solver in the message with the enclosing solver when used as a context manager.
-        msg["kwargs"]["solver"] = self
+from chirho.dynamical.internals.backend import Solver
 
 
 class TorchDiffEq(Solver):
